(PLINK) located wrapper classes into separate files.
